requirements.txt
azure-ai-documentintelligence==1.0.2
azure-core==1.38.2
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
isodate==0.7.2
python-dotenv==1.2.1
requests==2.32.5
# --- Phi-4 via Microsoft Foundry (uses requests) ---
typing_extensions==4.15.0
urllib3==2.6.3
